P:mw:maint:update_special_pages: updateSpecialPages in s6 to mw-cron
This pilots updateSpecialPages on a single shard. The choice of s6 is
somewhat arbitrary, but it is (joint with s8) the 2nd fastest shard to
run (~ 1h) that is not s1 (~ 35m) - i.e., pick the shard with minimum
aggregate "cost" that is not also a single (big) wiki.

Hosts: mwmaint1002.eqiad.wmnet, deploy1003.eqiad.wmnet
Bug: T388534
Change-Id: I36e8883ce5750b039f45b07bacd5e1256b3c77c9