autoinstall: setup for new apus nodes with boss card
These systems have their OS on the boss card (which does its own
RAID-1), so this is based on the standard flat.cfg. The changes are to
specify the use of /dev/nvme0n1 (which is how the boss card presents
to the OS), and to ensure any extant LVM is left alone (to prevent
reimage wiping Ceph data from the hdds).

Also, stop partman_early_command from attempting to set up any RAID on
these systems, as that would likely confuse the partitioner!

Bug: T392844
Bug: T392845
Change-Id: I9d8419563de0cb21dd056b0252929b287bde21ad