Deprecate AQS-related groups
These were only used/needed when the aqs Node service was co-hosted to
Cassandra directly on the aqs* servers.

This is no longer the case and the remaining Cassandra servers are managed
by SRE only.

As such, mark the groups are deprecated and remove the access to the servers.

Once the aqs* hosts are reimaged to Bookworm, we can remove aqs-roots
and aqs-admin and reclaim the GIDs

Hosts: deploy1003.eqiad.wmnet, aqs1017.eqiad.wmnet
Change-Id: Ic29da2b4c44e9fa2788b590267b314411e1b46ea