homer: make private repo support multiple peers
Why:
The current setup for the private repo supports only having 2 hosts
currently in production. When installing a new host it's useful to be
able to have a 3-hosts setup working.

What:
Convert the remote peer configuration to support multiple peers and add
the related remotes to the git clone setup. Convert one of the hooks
from a static file to a template to support the multiple peers.

Hosts: auto
Bug: T389380
Change-Id: I7b11af63481ed11c1de6a5e144d1ab6ece9ea6ba