alertmanager: Use alert summary as title for most tasks
These receivers are used for mw-cron, and the existing alert is too
generic. It creates one task for all the periodic job failures for
related to this project.
By using the alert summary as the title, one task per CronJob name will
be created, making the tasks more precise and easy to triage.

Bug: T385709
Change-Id: I1f899f10e97527a7c314a364dc72c1d7210d98e0