maps/osm: Move Kartotherian SQL config into Puppet
On the legacy maps servers, Kartotherian is co-hosted with the postgresql
stack. With the new setup this moved to Wikikube, but the imposm-initial-import
script still assumes a local checkout of the old Kartotherian as deployed
with scap.

As such, import the SQL statements into Puppet, install it on the maps
servers and update the imposm-initial-import to read them from there.

Hosts: maps-test2001.codfw.wmnet, maps1009.eqiad.wmnet
Bug: T381565
Change-Id: Iac4ec2d61b653b4b21b62120e49d379d06a9c243