maps/osm: Make create_layers_functions independent of Kartotherian checkout
On the legacy maps servers, Kartotherian is co-hosted with the postgresql
stack. With the new setup this moved to Wikikube, but the imposm-initial-import
script still assumes a local checkout of the old Kartotherian as deployed
with scap.

As such, import the SQL statements used by create_layers_functions into
Puppet and update it to read them from there.

Hosts: maps-test2001.codfw.wmnet, maps1009.eqiad.wmnet
Bug: T381565
Change-Id: I1cb15eeb3c989fc4bff208eca0eaf012f1f4fb18