cephadm: handle storage servers with BOSS card
osd_spec.yaml (uses rotational devices for data, and devices matching
NVMe for block.db) is applied to hosts labeled NVMe.

The new storage servers have the OS on a boss card, which appears as
nvme0n1 (and that ceph will ignore for OSD purposes as it already has
the OS on it), and then the NVME for OSD use as nvme1n1.

So where we find nvme0n1 is a BOSS and nvme1n1 is an NVMe then we can
apply the NVMe label. For the template to do this, though, we must
also add blockdevice_nvme1n1_model to the host_details hash we look up
from puppetdb.

Hosts: moss-be1001.eqiad.wmnet,moss-be2001.codfw.wmnet
Bug: T391354
Change-Id: If527957e68a4d5efdb6f009a034719ed90f1476c