Profile::Mediawiki_deployment: add 'clusters' field
The 'clusters' field is added in [0] and will be used to configure the
specific k8s cluster names relevant to a given deployment.

If none is provided, the default value is that from the 'k8s_clusters'
key in scap configuration (i.e., previous behavior).

Example: The mediawiki-dumps-legacy deployment exists in dse-k8s-eqiad,
so this will look like:

  - namespace: mediawiki-dumps-legacy
    clusters: ["dse-k8s-eqiad"]
    dir: "dse-k8s-services"
    [ ... snip ... ]

As before, the documentation header on Profile::Mediawiki_deployment
refers readers to the DeploymentsConfig docstring for an explanation of
these fields.

[0] https://gitlab.wikimedia.org/repos/releng/scap/-/merge_requests/795

Hosts: O:deployment_server::kubernetes
Bug: T388761
Change-Id: If4ac19cf6cc13d0b54467e03931bf28a934a14c4