mariadb: Change x2 and x3 ports to avoid conflicts with extra port
By default, extra port for standalone instance is 3307.
For multiinstance hosts, it is port + 20.

As x2 and x3 used the 333X and 334X range, respectivelly, they
conflicted with the sX and mX/x1 extra port section. Move them
to the 336X range, like this:

x2: 3330 -> 3362
x3: 3340 -> 3363

Comment that we should only use for now the 3310-2329 and
3350-3369 ranges for now (and skip every 20 ports).

Bug: T384274
Change-Id: I5e91bd786e3844c1a793f49e8f9348599c21a387