gerrit/nftables_throttling: make abusers more generic
This change moves the gerrit_abusers logic from profile::gerrit
to profile::firewall::nftables_throttling. This allows to drop
IPs or ranges on all nftables hosts, not just gerrit (like lists
or gitlab).

By default the list of abusers is empty, for Gerrit this should be
a NOOP beside some naming changes.

This is a workaround until BLOCKED_NETS can be used (T348734).

Bug: T394519
Bug: T348734
Change-Id: I8695f0d9938e9fe366910d0e09e8ac97317b8724