mw:maintenance: Adapt generatecaptcha for mw-cron
Why:
We need to be able to access the words and badwords lists from helmfile,
which means they need to be either in the chart (which is not possible
due to them being puppet secrets), or in a yaml value in
helmfile_defaults.

What:
Create a fancycaptcha.yaml values file that will be added to mw-cron's
helmfile for consumption by the chart, allowing us to create configmaps
to include in the mw-cron containers.

Hosts: mwmaint1002.eqiad.wmnet, deploy1003.eqiad.wmnet
Bug: T388531
Change-Id: If0f10634cb7a8db72f974098cb0bbe0f0e72ef84