zuul: add contint-roots admin group to new zuul::main role
This will give shell access with full root on the new zuul
"main" VMs to the members of the existing contint-root group
which is also used on contint servers.

Includes the assumption that James Blair will be added to it
as part of his access request which will give him both old
and existing prod CI server access.

Bug: T394819
Change-Id: I184f6a078898a02a0a44d3b308cc2b6e13d20256