profile::prometheus::k8s: drop terminated pod targets
k8s service discovery with role: pod will produce scrape targets for all
pods known to the API, including terminated pods.

While this is generally not a problem, as terminated pods tend to be
garbage collected quickly, this is not the case for pods associated with
k8s jobs, which will linger until ttlSecondsAfterFinished expires.

This leads to spurious metric collection when the pod IP associated with
the terminated job is reused, as metrics collected from the new pod will
be attributed to both the real and "zombie" targets. See T395052.

For configs that use role: pod in contexts that are broad enough to run
into this issue (i.e., not k8s-pods-istio), we now drop metrics from pod
targets in terminal phases.

Hosts: O:prometheus
Bug: T395052
Change-Id: I35d50a6cead62b5328cdaafed5a907f31984af77