admin/data: create an airflow-deployers group
That group will initially contain anyone having been added to the
analytics-privatedata-users group, and will be used to manage
permissions to deploy airflow and airflow development environments.

This will make sure data engineers don't need to be part of the
deployment group (which comes with permissions to deploy mediawiki
itself) to deploy a simple devenv.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T395125
Change-Id: I7df17bfa7e7c19971e8a1619e879ade8f1fe913c