site/zuul: create skeleton role/profile for new zuul executors/runners
Introduce new skeleton roles and profiles for new zuul machines
with the roles of "executor" and "trusted runner".

All they do is install docker for now.

But we need them to apply access groups and want to make clear in
site.pp what is used for what.

Even if everything happens inside a container they do have different
roles with different needs for firewalling/security/resources.

Bug: T394819
Bug: T393873
Change-Id: Ib7c3c7cc8ec2c8348fe95c3a8b2b90f193271659