Allow all users to traverse /etc/helmfile-defaults/private and subdirs
We need to be able to use the helmfile secret mechanism for users who
are not part of the 'deployment' group. Our first use case for this is
the on-demand Airflow development environments on the dse-k8s cluster,
where we do not wish all of the developers to have to be members of the
deployment group, but they need access to a specific secret.

At present, the /etc/helmfile-defaults/private and
/etc/helmfile-defaults/private/dse-k8s_services directories are
configured with mode 0750 and owned by `root:deployment`

This prevents users outside of the deployment group from traversing these
directories, even if they have access to the files contained beneath
them.

This commit changes the file mode of those directories to 0751 and
thereby permits all users to traverse the directories, but not to list
the files or access their contents.

Hosts: O:deployment_server::kubernetes
Bug: T393998
Bug: T395125
Change-Id: I3fadcdf2bafcdbf6addf27d42c8281558e030fa5