deployment:fix-staging-perm: update fix-staging-perms
Kubernetes hosts running mw-experimental will have a
copy of MediaWiki located at /srv/mediawiki. We want
deployers to have read/write access to this directory,
and be able to fix permissions if needed.

For simplicity, this updates the existing script, as it
will be used in Ida754afab2add9f48519703f59f57716a7e781f8

Bug: T276994
Change-Id: I89cf62a41893e8d426e8fdea80f8ed72ee47bbd0