doc: add support for PHP version bookworm, drop php_prefix variable
We need to make this profile work on bookworm.

For that we need to add support for a PHP8 version.

Per comment from Moritz on Ie1d64dedff456067, there is no reason
we have to use the same version that appservers use, as we originally
thought.

Instead we can just the default version in Debian, which is 8.2,
and for that we can just set it to wmflib::debian_php_version().

Do this without affecting existing servers on bullseye.

Additionally, the $php_prefix variable is simply the string
php followed by the version.

So simplify this and just drop the separate prefix variable entirely.

Bug: T392130
Change-Id: I5ca52ed103ace2ec49a59cc793734badaad55685