mediawiki-dumps-legacy: Remove user:group overrides for k8s config
We now wish to use scap to deploy the mediawiki-dumps-legacy/production
release to the dse-k8s cluster.

In order for that to work, we need to ensure that scap can read the k8s
config files, which have been installed with access to the
analytics-deployers group.

Hosts: O:deployment_server::kubernetes
Bug: T389786
Change-Id: Ie00d10c90a9df718b87c0e3d84416ceda9ca0c46