Allow an-druid1003 to reformat its data drives
We need to reimage an-druid1003 due to a RAID drive failure on the drive
with grub installed.

For this reason, we need to revert the preseed configuration to allow
the host to reformat all drives.

We are not concerned with losing data, since the data on these drives is
ephemeral. It will pull it again from HDFS.

Bug: T393229
Change-Id: I53abdbcdc6ac4190628b61e69e21f8872244efe6