cache::haproxy: start optimizing for readability
As of now, the configuration of the tls terminator is extremely flexible,
but it is not very readable or easy to understand. This commit starts by
making parts of the configuration that are either:
* Common to all deployments
* Part of a specific feature that already has a feature flag
* Should be part of a feature flag and/or the default configuration

explicit in the configuration. This will make it easier to understand.

Hosts: cp4050.ulsfo.wmnet,cp4038.ulsfo.wmnet,cp3066.esams.wmnet,cp3081.esams.wmnet
Change-Id: Ic191ace3432206099b6e2d7a7b525d15414d0e9e