zuul: create profile to setup system user and group
As mentioned by Dan Duvall in our zuul-meeting, we
will need a puppet setup similar to the existing one
for buildkitd, with a systemd service that runs
a docker image. (modules/buildkitd/manifests/init.pp)

While most of it remains TBD it seems already clear
we will definitely need a system user and group for that.

So starting to create that and putting it in a separate
profile so we can include it in multiple roles without
repeating ourselves.

This system user needs to be a member of the docker group.

The docker group already exists as it was created by the
docker package that is installed by puppet already.

Change-Id: I9c377b927c39521b24d525c6c4672e78b745beb9