Revert a bunch of vendordata changes
Some of these things worked around the problem, but
the real issue was that the subnets were configured to
use an invalid dns server. With that corrected we should
be able to go back to our original setup without
systemd-resolved.

The actual fix for all this was

https: //gitlab.wikimedia.org/repos/cloud/cloud-vps/tofu-infra/-/merge_requests/238

Change-Id: I66f25eaa8f80b61906ad9bc0bace2205ffca8e82