varnish: Replace X-Include-PV with include_pv var
This one's a little weird since there's a control flow test on whether
include_pv is 0 but var.get_int() defaults to returning a 0 on an
undefined variable. This could be a pitfall, causing truthiness as True
even though it's 0.

To avoid this confusion, just use a string - var.get() defaults to ""
on a nonexistent var and is much clearer to understand.

Bug: T373550
Change-Id: I9ec4aef570865f20a6921e83680ca2b94608c596