backup: add doc1004 to job_monitoring_ignorelist
As part of T392130 doc1004 is a new host replacing old host
doc1003 (along doc2003 replacing doc2002).

Temporary ignore the backup monitoring because otherwise
it alerts when the production puppet role is applied for
the first time.

Avoid monitoring noise for Jaime as the owner of the backup servers.

I got this string with:

- ssh backup1001
- sudo bconsole
- list jobs client=doc1004.eqiad.wmnet-fd

Bug: T392130
Change-Id: I2f413cf85aa80bb4786ef7905de5d1726fd99ac3