scap: block interactive maintenance scripts on mwmaint
In advance of decommissioning the maintenance hosts, block interactive
mwscript and foreachwikiindblist runs on mwmaint hosts.

This strengthens the language in the deprecation banner in mwscript and
replaces the sleep with exit 1. In the case of a script that has no k8s
solution at the moment, users are directed to the deployment host.

In addition, while the above takes care of mwscriptwikiset as well, it
seems the banner was never added to foreachwikiindblist. This now does
so (which also takes care of foreachwiki).

Bug: T341553
Change-Id: I580d622d6728eb9be3ccbce72fcd9c57a5c206c5