logstash: add early-stage filter to populate event.original
Why: From the ecs docs, event.original is intended to to demonstrate
log integrity and is needed if reindexing is to be done.  In addition,
the loss of data is not optimal, yet common due to indexing failures
and schema incompatibilities.

What: Unless provided by producers or ingest daemons, instruct
Logstash to generate the event.original field from the event data it
gets from Kafka before any transformations are applied.

Hosts: logstash1023.eqiad.wmnet
Bug: T234565
Change-Id: I1761f74bf4f2ccff9719095f319d811d84d5bc12