deployment_server: Update the local helm cache in mwscript-k8s
Previously, the fallback to using a local helm cache in the restricted
user case only initialized the cache if it did not exist.

However, there was no mechanism to _update_ the cached repository index,
so users would end up with a one-time snapshot of the index that drifts
out of date (and thus use stale chart versions).

With this change, restricted users will now refresh their cache on each
mwscript-k8s run. In practice, this is a rather quick operation, so it
should not pose a significant usability issue.

Bug: T395521
Change-Id: I48153dd9f8c4429a0861eb5d9bb51fdaf382a899