haproxy: remove conditionals on wikimedia_trust
The stanza is always defined as it's not just a parameter. Even if it
was, there's plenty of other places in the config where the ACL is
assumed to exist.

So, rather than checking if the variable exists here, we do so in puppet
where we can set a default instead.

Change-Id: I1eab5e73d4d6f14d3002d2fcd177d306b08a44b3