Revert^2 "gerrit: add a second replica, start replicating to gerrit2003"
This reverts commit d3e2ea3668491ec8a37cb3478272b61fcb56c2a7.

Reason for revert: wrong user name fixed and just needed service restart

In addition to the revert this is now harcoding user name "gerrit"
for host gerrit2003.

Once all hosts have been reimaged it will just be "gerrit" instead
of "gerrit2" for all of them and we can either change the variable
to that or just hardcode it forever for all.

Bug: T395887
Bug: T338470
Change-Id: Ica961c0eb07b58dcb35946125f9963bc58144e31