logstash: Reroute apache.access logs to webrequest partition
Why: Apache access logs are over half of the events in the ecs-default
partition.  This is to reduce the overall size of the ecs-default
indexes and follow the pattern of the legacy (logstash-webrequest)
indexes.

What: Route apache.access logs to the ecs-webrequest partition and
amend curator config to handle the new indexes.

Bug: T390215
Change-Id: I005dfc562bbcc1eb4f11e1650f2d62ecb837faf1