tlsproxy: Specify a default retry_on policy
Why:
Issues continue to persist in T380958. Re-reading the docs, it appears
that specifying an explicit retry_on stanza for the retry_on for the
retry_policy is required.

What:
Specify that in case of connect-failure, which isn't clear from the docs
whether it covers UpstreamConnectionTermination causes, but probably
does, 1 retry should happen by default

Hosts: auto
Bug: T380958
Change-Id: I1c9a734e517bcc0aa0f85f040a92e22f2c61586c