liberica: Don't bind VIPs to lo interface with katran
With katran adding the VIPs to the loopback interface is no longer
required. XDP will capture the traffic headed to the VIPs before the
regular networking stack has the chance of dropping the packets.
Additionally this interferes with healthchecks targeting the VIP:
If the VIP is present in the routing table of the load balancer,
the source IP is set to the VIP and the realserver response to the
healthcheck is unable to reach the loadbalancer.

Hosts: lvs6001.drmrs.wmnet
Bug: T395228
Change-Id: I01b3eec23245a6b768042ff06a1e2a3f0242dae1