tlsproxy: Switch retry_on to 5xx
Why:
connect-failure, despite sounding like it would solve the issue, did
not. 5xx includes connect-failure and refused-stream. In a series of
repeated siege tests (the command is in T380958) against mwdebug2001
(with connect-failure) and mwdebug2002 (with 5xx manually set) we
repeatedly saw 1 or 2 errors for mwdebug2001 and 0 errors for
mwdebug2002. Thus, it makes sense to try out 5xx

What:
Switch retry_on from connect-failure to 5xx

Hosts: auto
Bug: T380958
Change-Id: I00af6ac572648168a79c46fd6da7498dd72af4dc