atftpd: Add support for Bookworm
Since we can only install VMs in routed Ganeti with Bookworm and later
(the netcfg changes were only backported to Bookworm), we'll need to
also install the new magru install server with Bookworm.

This adapts atftpd, the major change is that it now uses socket activation
instead of bespoke daemon code.

And rather then fixing it, let's remove the spec file, the remaining
use of that test is whether a file is present, which is hardly useful...

Hosts: install3003.wikimedia.org,install7002.wikimedia.org
Bug: T396487
Bug: T394263
Change-Id: I58c19633bfea682916da9901fa1a00e1c1470ca1