Add radosgw access for members of the new 'object_storage' role.
If a user has this role without the 'reader' role they won't
be able to access the Horizon UI but should be able to use APIs
and CLIs. I user with 'reader' and 'object_storage' but not 'member'
will be able to use the swift UI in Horizon without being able to edit
VMs &c.

Bug: T396594
Change-Id: I124d8b86922e910c10888eab9c48b9739a01cc0c