tables-catalog: Add Flow tables
Flow has two configurations:
* For public wikis, the tables exist in `flowdb` on `x1`, they are
  shared between all of the wikis where it's enabled.
* For fishbowl and private wikis, the tables exist in each wiki's
  default database (only such wiki where it's enabled is officewiki,
  see flow.dblist).

Additional tables `flow_definition` and `flow_wiki_ref_deleteme`
currently exist in `flowdb` on `x1`, they are not referenced by the
code and should be dropped.

I marked all tables as "canonical" (out of abundance of caution) and
"private" (as they are not currently replicated). I don't think it's
worth the effort to find out if this could be changed.

Bug: T363581
Change-Id: I50689de23e18204c0da7276bda763adef5a713b4