tables-catalog: Add LiquidThreads tables
LiquidThreads tables exist in each wiki's default database.

It's enabled on wikis that set $wmgUseLiquidThreads or
$wmgLiquidThreadsFrozen (there is no dblist, just config variables).

The `historical_thread` table is unused and should be dropped,
T396675.

Some or all of its tables currently exist for some other wikis where
it has already been undeployed (in some cases years ago), they should
be dropped.

I marked all tables as "canonical" (out of abundance of caution) and
"private" (as they are not currently replicated). I don't think it's
worth the effort to find out if this could be changed.

Bug: T363581
Change-Id: I39762fbc1f7562f3a98138e61095e32e3da06e5b