microsites: adjust monitoring for os_reports to k8s
Since the os-reports.wikimedia.org site moved from legacy miscweb
to kubernetes, adjust the monitoring accordingly by:

- adding Kubernetes Ingress port instead of default 443
- adding Kubernetes ingress v4 IP address
- adding parameter to set certificate expiry to 9 days

This is all copied from other existing monitoring checks for
services we moved to Kubernetes.

Additionally, move the code from the profile specific to os_reports
to the same place with the other monitoring checks.

The next step will be to move all of microsites::monitoring to a new
place before decom'ing the legacy VMs.

Bug: T397080
Bug: T350794
Change-Id: I22a20acb00532895beffea10e98e1e958d379859