zuul::executor: add zuul user and nodepool ssh private key
Add the system user and group zuul on the zuul executor nodes.

Create /etc/zuul/ and /etc/zuul/ssh.

Add the newly created SSH private key that has been added to the
secrets module in the private repo.

This is the "global nodepool ssh key" for executors to connect to
workers.

Later the public part has to be added on workers in cloud VPS.

Bug: T395938
Change-Id: If9cd93d76efe7ea1518981bc7caa0c8d7c0ca9d7