Add UploadWizard tables
Notes:

I excluded `uw_campaign_conf` because the only place
it's referenced in the code is in a maint script to "migrate
campaigns from older, database table to newer page based storage"

I included all dbs that I can find with the table uw_campaigns
even though some of them (e.g. donatewiki) don't contain any
data

uw_campaigns table defn has a note that says "This is not the
primary storage for campaigns". The db table gets populated when
a campaign page is saved, but if the data is lost then
Special:Campaigns won't work until each campaign page is re-
edited, so marked it as canonical

Marked visibility as private because the table is not visible
in commonswiki_p

Bug: T393793
Change-Id: Ic0f0f0040bf1f40ddcba73f96dc97f48343c46f0