fetch_external_cloud: stop depending on requestctl libraries
We have moved the requestctl code to the HIDDENPARMA repository with the
goal of allowing a faster pace of development and not lock the fate of
the rest of conftool with requestctl itself.

To complete this work, we need to remove the dependencies from
the rest of our codebase, and switch to using the command-line tool.

Change-Id: Ic34f8304f9a4aa77e6ae1897cd2c0a3160363985