deployment_server: use bookworm httpd in mw-debug/next mw-*/migration
Requires that make-container-image is updated with [0].

This switches mw-debug/next and all migration releases to the bookworm
webserver image flavour.

This will allow us to manually verify the new image in mw-debug, and
then pilot a bit of traffic on all large mediawiki services.

Note that, as of now, all migration releases are scaled to zero.

[0] https://gitlab.wikimedia.org/repos/releng/release/-/merge_requests/184

Bug: T378128
Change-Id: Iaf1f4ad16363c09b165ac280ef391339387773a7