Neutron policy: Update port policies
Because we use a shared network among multiple projects,
we need slightly more permissive rules here. I've tried to
forward this upstream but without success

https://review.opendev.org/c/openstack/neutron/+/886214

This change is meant to leave the shared network behavior the
same while updating our rules to resemble upstream changes.

Two rules have been removed as they're now identical to
the upstream rules.

The create_port rule has been removed as the upstream rule
is better; it allows the service_api access which we need
for nova to work properly as a service user.

Change-Id: I73f8698ef8f7e52efa8aa927b98608b6b4d12321