Add new reposync repo called 'netbox-dns'
This will create a new repo on the hosts enabled for reposync
(cumin and netbox hosts) called 'netbox-dns'.  Ultimately we will be
writing files to this from a new cookbook which generates DNS records
based on the names on IP objects in Netbox, and eventually syncing the
contents of the repo to our authdns boxes and migrating from our
current 'dns snippet' repo/directory.

Work being carried out as part of the I/F hackathon.

Bug: T362985
Change-Id: I96afb4ba947c0206750b2b80fba8763fc0eb508c