Netbox hosts: ensure reposync repos are set up to match cumin hosts
The cumin hosts are configured to set up reposync repos with names
matching the items under profile::spicerack::reposync::repos.

It sets the netbox hosts as remotes for these as they provide the
git front-end web server.  Given that is the case it seems to make
sense to also set up the repos on the netbox server based on the
same key, so they are always in sync.

This is an alternate approach to that in
I20fe7a6157ac0172a9a0d02c902a733f91e87fed and may be preferable.

We had to move the list of reposync repos used by the
cluster::management role to the less-specific spicerack profile level
so it is readable in profile::netbox::automation, which was otherwise
just getting the empty array every time.  As this resource is only
used in one place this should be safe.

Hosts: netbox1003.eqiad.wmnet, netbox2003.codfw.wmnet
Hosts: cumin1002.eqiad.wmnet, cumin1003.eqiad.wmnet
Hosts: cumin2002.codfw.wmnet
Bug: T362985
Change-Id: Ia663266a248b3a7a3fa5b076edbc91b69edbdba7