hdfs: set an-worker1176 to analytics-fex recipe
During the disk swap, the host was observed to have some I/O errors affecting the root partition.
After a reboot the host can no longer boot correctly as the root partition is no longer available,
The host is not part of the hadoop cluster just yet and is safe to reimage with the original recipe.

Bug: T390176
Change-Id: Iafaa7213b322833450076a0560223cacdb760a09