Authdns: clone new netbox-generated DNS records repo
This patch will cause our dns hosts to clone the new git repo we
are saving DNS records generated by the sre.dns.netbox-records
cookbook.  The repo will be cloned to /srv/git/netbox_dns_records,
with the idea that the files in that dir will eventually be used
instead of those produced by the legacy script in
/srv/git/netbox_dns_snippets.

For now the only thing this does is to create the directory for the
new repo and sync it to the hosts, nothing else changes and dns
operations will not be affected.

Hosts: dns3003.wikimedia.org, dns1004.wikimedia.org
Bug: T362985
Change-Id: Ie7dd6e4bff116b7f94aaf08c78cb035ed88bd96f