os_updates: manage stylesheet with puppet
After mounting a volume to allow the
rsync to auto-update the original files
in the document root were "hidden". This
seems like the most straightforward way
to fix this. It places the CSS file in the
same directory as the generated html files
so they will all be fetched by the rsync

Bug: T350794
Change-Id: I90f8390c6f5c753f5b9e02e8b2601fcfb0044c08