puppetserver: check for rebase in puppetserver-deploy-code
Attempt to detect an active rebase process by looking for
a .git/rebase-apply or .git/rebase-merge directory. If found explain
that the puppetserver update will be skipped because of the active
rebase. This will hopefully be more informative and less alarming than
the prior logging triggered by the same underlying actions.

Bug: T397877
Change-Id: I99b1b4a32b6d2b942e276002a2d6d6db40dbd5d6