Revert "centrallog: Add a temporary rsyslog debug config file"
This reverts commit 6d332f14ffecb3a3ef82286fc948377149c93139.

Reason for revert: The change is not working as expected, it's only logging for
the host itself, I think I need to increase the rsyslog priority to process the
logs before they're routed to the host files.

While I work on a new patch I'll revert this one so we're not logging irrelevant
data.

Change-Id: I36d695673146780da794ae5c9b57a5f4a6ee64fc