Dumps_v1: Disable the sync job that publishes from dumpsdata servers
We are migrating the dumps from the snapshot and dumpsdata servers to
airflow and kubernetes.

This commit constitutes part of the migration plan, whereby the existing
dumpsdata servers stop running the job that publishes the dumps to the
clouddumps servers.

This should not be merged until late on June 30th.

There will be a corresponding change to airflow-dags that updates the
destination directory of the dumps sync tasks to the production target.

Hosts: C:dumps::generation::server::rsyncer_common
Bug: T397848
Change-Id: If2ce6e3b66db4614c2a87239a15ea8ed1d56e75a