Dumps_v1: Stop updating dumps monitor HTML/JSON from the legacy system
We are migrating the dumps_v1 from the snapshot and dumpsdata servers to
airflow and kubernetes.

Part of this work includes updating the way that the HTML and JSON
status pages for the dumps are produced and published.

As part of the migration, we need to enable the new method, which is an
airflow DAG, and disable this legacy method, which runs as a timer on
the clouddumps and fallback dumpsdata servers.

Hosts: C:dumps::web::dumpstatusfiles
Bug: T397848
Change-Id: Ifce8073bc8cd003d07137683f41d048e164fe44b