deployment_server: use bookworm httpd in all mediawiki releases
Now that ~ 48h have passed with mw-api-ext and mw-web serving 5% of
traffic on the new image, and no issues discovered, migrate all traffic
and releases.

This continues to use the webserver-bookworm flavour rather than simply
rebasing webserver on the bookworm image. We will do that, but not until
a later step, since keeping them distinct makes rollbacks easy.

Bug: T378128
Change-Id: I835608ecdea82e72a16c5330d7b6a0a11155920d