P:etcd::tlsproxy: fix notify behavior for PKI certs
In I0c6aafa1b0bbbd1d54148d1c777ede3c05416d32, there were two closely
related issues with how notify behavior was configured:
* It notified Server[nginx] on updates to certificate material, which
  manifests as a *restart* rather than reload.
* Since this resource differs from what's used on nginx config updates
  to trigger proper reloads, the refreshes could not coalesce.

Together, this gave us a reload and a restart, leading to the race we
saw during the pilot migration on conf2006.

In any case, the behavior we want is to notify Exec[nginx-reload]. We
can achieve that in a couple of ways:
1. Configure a notify on cfssl::cert, which will propagate to all
   contained resources, including the ones we care about.
2. Add more plumbing to cfssl::cert that parallels notify_services.
3. Introduce a notify relationship externally (~>) that targets the
   specific resources we're interested in.

Adding more knobs to cfssl::cert (#2) seems non-ideal. I considered #3,
but:
* We should not depend on the names of the exec resources internal to
  cfssl::cert.
* I'm not convinced File[(cert file)] ~> Exec[nginx-reload] would work,
  since contents are managed by exec resources (not the file resources).

Given that, I opted for #1. While quite broad, in that the notify will
propagate to all contained resources, I suspect it's not going to make a
great deal of difference in practice. See e.g. discussion on [0].

[0] I2d78250f05d32a623510d63eb91d6ca8ea5aaf9e

Hosts: O:configcluster
Bug: T352245
Change-Id: I6e0263a911f3944e67cd3304bf45fc47175b314a