Ensure that master=yarn is the default spark configuration for users
We have become aware that sometimes users omit the '--master=yarn'
command-line argument when starting a spark session. This can have the
unintended result of running all execution on a single machine, rather
than distributed across the Hadoop cluster.

In order to tackle this, we add the default spark.master configuration
option to the spark3-defaults.conf file.

Users can still choose to override this on the command-line to use a
local master if they wish, but it would have to be a more deliberate
choice to do so, rather than an omission.

Hosts: P:hadoop::spark3
Bug: T393181
Change-Id: I17227773485838f4a4c167e7c42e469d56ab92ae