cache,haproxy: refactor haproxy captures to fix x-analytics logging
Our current HAProxy configuration is based on declaring capturing
slots for headers and vars and then log these when needed, during the
request, the response or after the response, depending on the kind of
information we need to access.

Haproxy provides different ways to declare and capture, this change
tries to address the different syntaxes and consolidate into one way
to declare and log capture slots (at least for response, requests will
come in a separate commit).

This change also address the issue related to the requestctl rules
that aren't sent to Varnish (managed entirely in HAProxy) and not
logged. The work started with
I17564088dafa56b1b6206b92e6442d698e374fe5 was not enough because we
need to capture the header *after* it was modified in the requestctl
backend, so using the http-after-response capture directive looks like
the only way to log this information.

As the X-Analytics header is removed before being sent to client, the
safest way to log it was to copy into a dedicated variable and modify
the log format to access that instead of the response header. The
header is also "sanitized" from the requestctl part in case of
X-Mediawiki-Debug.

We also removed the (unused since long time) extended_logging flag as
haproxykafka is now deployed everywhere.

Lastly, configuration has been reordered and some comments added for
better readability

Hosts: cp1111.eqiad.wmnet,cp7001.magru.wmnet
Bug: T397917
Change-Id: I2a1eb461d96ebbe2e3a64eaac538023a13f65b60