prometheus: add dnsbox_service_state_exporter
On the DNS hosts, we have no visibility into the state of services as
controlled by confd/confctl. This means that we can consider a service
pooled in confctl but if the healthcheck is failing, anycast-hc will not
set the prefixes file for bird, which means bird will not advertise the
prefixes (for that service).

anycast-hc exports Prom metrics but it can't know what the state of the
service as we set it from confctl. To fix that, start exporting metrics
for the state of the service from the DNS hosts. This will allow us to
have historical data and we will also use this for the work in T374619
below to alert if there is a mismatch between the service state we set
and as bird advertises.

Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org
Bug: T374619
Change-Id: I1c21682f998627433f1f8d39a7a74d6c635e7fba