docker_registry::web: allow prod-build to pull restricted images
The build nodes are running a new version of docker-report that
is able to inspect all the images running in containers on various
k8s clusters, to then run the debmonitor client on them.

It would be great to start inspecting the /restricted/ images too,
and as far as I can see the 'prod-build' user needs to be allow to
pull (and indeed I see HTTP 401 errors as a consequence on build nodes).
If the prod-build user is too broad (since it may include also CI),
we can possibly create a new one just for this use case.

Add build2002's IP to the list of the host allowed to push.

Bug: T404437
Hosts: auto
Change-Id: I8772a9b2693603a90952ae7611bc70c1387642f0