deployment_server: group chown all airflow private files to airflow-deployers
Now that `airflow-devenv` requires reading the mirrored instance
private data, to automatically infer extra connection blocks,
we need to grant access to the airflow private files to the
`airflow-deployers` group.

That group is defined as:
```
  airflow-deployers:
    gid: 714
    description: Base groups for all users deploying airflow instances
    members: [*analytics_privatedata_users]
```
which is better scoped that the `deployment` group itself, and would
allow engineers such as Andrew MCAllister to deploy a dev instance
mirroring airflow-wmde.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T393998
Hosts: deploy1003.eqiad.wmnet
Change-Id: Iad8acde4fb6352d1de5f20d37c33fe3519019024