wdqs: improve blackbox check
Our current blackbox either does not trigger, or takes too long to trigger.
As a result, our response to broken hosts is delayed and the user experience suffers.

Try using the /readiness-probe that pybal uses (which appears to work)
to see if we have better results.

Hosts: prometheus1005.eqiad.wmnet,wdqs1013.eqiad.wmnet
Bug: T398341
Change-Id: Ib905af6eb18545069f76f9a9ecc39db8d026a7db