gerrit: config replicas for rename-project plugin
When renaming a project for the first time (T365985), I have noticed the
repository was not renamed on the replicas.  The reason is the
rename-project needs to be explicitly configured (it does not borrow the
configuration from the replication-plugin).

The renaming is done over ssh using the `gerrit rename-project` command
on each of the replicas Gerrit daemon. To do so, allow traffic from the
primary host to port 29418 (we only allowed traffic to the service IP).

Configure the rename-project with a hardcoded list of replicas.

Bug: T239693
Bug: T398401
Hosts: O:gerrit
Change-Id: I97fe0719da5923542ae0add11a0831e9e728d157