pyrra: remove multi-dc for istio-based SLOs
For istio services it doesn't make a lot of sense to show
multi-dc slos to the users, that may be confusing to look at
(if I were an SLO owner I'd expect only one).

Add a regex for the site label that matches the list of DCs
passed to the "datacenters" variable.

Bug: T398534
Change-Id: I3e39fdd651f021b84a7489cb862ebd73252eecca