openstack::patch: Disable fuzzing patch locations
By defualt, the `patch` utility will try to find similar locations where
the patch could be applied if the line numbers don't match exactly. This
is breaking the 'unless' command for the trust schema patch currently
applied to Keystone, where it found an another property that sort of
matched and thought the patch had indeed already been applied when it
was not.

Change-Id: I6a6a69643a12b7195d12e3410990453a14af3cb5