analytics: Absent rsync scripts that import Dumps 1 XML into HDFS
In this first patch for T396031, we want to absent files that import the XML generated by
the Dumps 1 process, and also stop doing data purge for the analytics tables that were
generated from these dumps. On a later patch we will remove the code itself.

Hosts: an-launcher1002.eqiad.wmnet
Bug: T396031
Change-Id: I778547039f029eab8eccbc78cf19e5f9ea076b95